Refactor piece creation
Rather than rasterizing each puzzle piece to an image and uploading to
the GPU, we now instead translate the bezier curves into line segments,
and tesselate the shape using earcut.

This allows us to upload *just* the source image to the GPU, and allows
the geometry to be perfect at any zoom level. As a side bonus, piece
edges are no longer aliased, and connections are completely seamless.